Refactor `commodore.dependency_mgmt` into a package
