+++ tests/test_dependency_mgmt.py
@@ -3,7 +3,6 @@ Unit-tests for dependency management
 """
 from __future__ import annotations
 
-
 import os
 import click
 import git
@@ -110,164 +109,6 @@ def test_create_component_symlinks(capsys, data: Config, tmp_path):
     assert capsys.readouterr().out == ""
 
 
-@pytest.mark.parametrize(
-    "libaliases,expected_paths,stdout",
-    [
-        (None, [], ""),
-        ({}, [], ""),
-        (
-            {"foo.libsonnet": "my-component.libjsonnet"},
-            ["dependencies/lib/foo.libsonnet"],
-            "",
-        ),
-        (
-            {"foo.libsonnet": "bar.libjsonnet"},
-            [],
-            " > [WARN] 'my-component' template library alias 'foo.libsonnet' "
-            + "refers to nonexistent template library 'bar.libjsonnet'",
-        ),
-        (
-            {
-                "foo.libsonnet": "my-component.libjsonnet",
-                "bar.libsonnet": "my-component.libsonnet",
-            },
-            ["dependencies/lib/foo.libsonnet"],
-            " > [WARN] 'my-component' template library alias 'bar.libsonnet' "
-            + "refers to nonexistent template library 'my-component.libsonnet'",
-        ),
-    ],
-)
-def test_create_component_library_aliases_single_component(
-    capsys,
-    tmp_path: Path,
-    data: Config,
-    libaliases: Optional[dict],
-    expected_paths: Iterable[str],
-    stdout: str,
-):
-    component = setup_mock_component(tmp_path)
-    data.register_component(component)
-    inv = Inventory(work_dir=tmp_path)
-    inv.ensure_dirs()
-
-    cluster_params = {
-        component.parameters_key: {},
-        "components": {
-            component.name: {
-                "url": f"https://example.com/{component.name}.git",
-                "version": "master",
-            }
-        },
-    }
-    if libaliases is not None:
-        cluster_params[component.parameters_key] = {
-            "_metadata": {
-                "library_aliases": libaliases,
-            },
-        }
-
-    dependency_mgmt.create_component_library_aliases(data, cluster_params)
-
-    expected_aliases = [(tmp_path / path, "lib") for path in expected_paths]
-    for path, marker in expected_aliases:
-        # Ensure symlinks exist
-        assert path.is_symlink()
-        # Ensure symlink targets exist
-        assert path.resolve().is_file()
-        # Ensure symlinked file contains correct marker content
-        with open(path) as f:
-            fcontents = f.readlines()
-            assert fcontents[0] == marker
-
-    captured = capsys.readouterr()
-    assert stdout in captured.out
-
-
-@pytest.mark.parametrize(
-    "tc1_libalias,tc2_libalias,tc3_libalias,err",
-    [
-        ({}, {}, {}, None),
-        (
-            {"foo.libsonnet": "tc1.libjsonnet"},
-            {},
-            {},
-            None,
-        ),
-        (
-            {},
-            {"foo.libsonnet": "tc2.libjsonnet"},
-            {},
-            None,
-        ),
-        (
-            {"foo.libsonnet": "tc1.libjsonnet"},
-            {"foo.libsonnet": "tc2.libjsonnet"},
-            {},
-            "Components 'tc1' and 'tc2' both define component library alias 'foo.libsonnet'",
-        ),
-        (
-            {"foo.libsonnet": "tc1.libjsonnet"},
-            {"foo.libsonnet": "tc2.libjsonnet"},
-            {"foo.libsonnet": "tc3.libjsonnet"},
-            "Components 'tc1', 'tc2', and 'tc3' all define component library alias 'foo.libsonnet'",
-        ),
-        (
-            {"tc2-fake.libsonnet": "tc1.libjsonnet"},
-            {},
-            {},
-            "Invalid alias prefix 'tc2' for template library alias of component 'tc1'",
-        ),
-    ],
-)
-def test_create_component_library_aliases_multiple_component(
-    tmp_path: Path,
-    data: Config,
-    tc1_libalias: dict[str, str],
-    tc2_libalias: dict[str, str],
-    tc3_libalias: dict[str, str],
-    err: Optional[str],
-):
-    c1 = setup_mock_component(tmp_path, name="tc1")
-    c2 = setup_mock_component(tmp_path, name="tc2")
-    c3 = setup_mock_component(tmp_path, name="tc3")
-
-    data.register_component(c1)
-    data.register_component(c2)
-    data.register_component(c3)
-
-    cluster_params = {
-        c1.parameters_key: {
-            "_metadata": {"library_aliases": tc1_libalias},
-        },
-        c2.parameters_key: {
-            "_metadata": {"library_aliases": tc2_libalias},
-        },
-        c3.parameters_key: {
-            "_metadata": {"library_aliases": tc3_libalias},
-        },
-        "components": {
-            "tc1": {
-                "url": "https://example.com/tc1.git",
-                "version": "master",
-            },
-            "tc2": {
-                "url": "https://example.com/tc2.git",
-                "version": "master",
-            },
-            "tc3": {
-                "url": "https://example.com/tc3.git",
-                "version": "master",
-            },
-        },
-    }
-
-    if err:
-        with pytest.raises(click.ClickException) as e:
-            dependency_mgmt.create_component_library_aliases(data, cluster_params)
-
-        assert err in str(e.value)
-
-
 def _setup_mock_inventory(patch_inventory, aliases={}, omit_version=False):
     components = {
         "test-component": {
@@ -307,165 +148,6 @@ def _setup_mock_inventory(patch_inventory, aliases={}, omit_version=False):
     patch_inventory.side_effect = inv
 
     return mock_inventory["nodes"]["cluster"]["parameters"]["components"]
-
-
-@patch.object(dependency_mgmt, "kapitan_inventory")
-def test_read_components(patch_inventory, data: Config):
-    components = _setup_mock_inventory(patch_inventory)
-    component_urls, component_versions = dependency_mgmt._read_components(
-        data, ["test-component"]
-    )
-
-    # check that exactly 'test-component' is discovered
-    assert {"test-component"} == set(component_urls.keys())
-    assert components["test-component"]["url"] == component_urls["test-component"]
-    assert (
-        components["test-component"]["version"] == component_versions["test-component"]
-    )
-
-
-@patch.object(dependency_mgmt, "kapitan_inventory")
-def test_read_components_multiple(patch_inventory, data: Config):
-    components = _setup_mock_inventory(patch_inventory)
-    component_urls, component_versions = dependency_mgmt._read_components(
-        data, components.keys()
-    )
-    # check that exactly 'test-component' is discovered
-    assert set(components.keys()) == set(component_urls.keys())
-    assert set(components.keys()) == set(component_versions.keys())
-    assert all(components[cn]["url"] == component_urls[cn] for cn in components.keys())
-    assert all(
-        components[cn].get("version", None) == component_versions[cn]
-        for cn in components.keys()
-    )
-
-
-@patch("commodore.dependency_mgmt.kapitan_inventory")
-def test_read_components_exception(
-    patch_inventory, data: Config, tmp_path: Path, capsys
-):
-    components = _setup_mock_inventory(patch_inventory, omit_version=True)
-
-    with pytest.raises(click.ClickException) as e:
-        _ = dependency_mgmt._read_components(data, components.keys())
-
-    assert "Component 'other-component' doesn't have a version specified" in str(
-        e.value
-    )
-
-
-@pytest.mark.parametrize(
-    "components,ckeys,exctext",
-    [
-        ({}, [], "Component list ('parameters.components') missing"),
-        (
-            {"components": {"a": {"url": "a_url"}}},
-            ["b"],
-            "Unknown component 'b'. Please add it to 'parameters.components'",
-        ),
-        (
-            {"components": {"a": {"version": "a_version"}}},
-            ["a"],
-            "No url for component 'a' configured",
-        ),
-    ],
-)
-@patch("commodore.dependency_mgmt.kapitan_inventory")
-def test_read_components_exc(
-    patch_inventory,
-    data: Config,
-    tmp_path: Path,
-    capsys,
-    components,
-    ckeys,
-    exctext,
-):
-    patch_inventory.return_value = {
-        data.inventory.bootstrap_target: {"parameters": components},
-    }
-
-    with pytest.raises(click.ClickException) as exc_info:
-        _ = dependency_mgmt._read_components(data, ckeys)
-
-    assert exc_info.value.args[0] == exctext
-
-
-@patch.object(dependency_mgmt, "kapitan_inventory")
-def test_discover_components(patch_inventory, data: Config):
-    component_inv = _setup_mock_inventory(patch_inventory)
-
-    components, aliases = dependency_mgmt._discover_components(data)
-    assert components == sorted(component_inv.keys())
-    assert sorted(aliases.keys()) == components
-    assert all(k == v for k, v in aliases.items())
-
-
-@patch.object(dependency_mgmt, "kapitan_inventory")
-def test_discover_components_aliases(patch_inventory, data: Config):
-    expected_aliases = {"other-component": "aliased"}
-    component_inv = _setup_mock_inventory(patch_inventory, expected_aliases)
-
-    components, aliases = dependency_mgmt._discover_components(data)
-    assert components == sorted(component_inv.keys())
-    assert set(components + list(expected_aliases.values())) == set(aliases.keys())
-    assert set(aliases.values()) == set(components)
-    assert aliases["aliased"] == "other-component"
-
-
-@pytest.mark.parametrize(
-    "components,expected",
-    [
-        ([], ""),
-        (["a"], "'a'"),
-        (["a", "b"], "'a' and 'b'"),
-        # Verify that Oxford comma is used in lists with >= items
-        (
-            ["a", "b", "c"],
-            "'a', 'b', and 'c'",
-        ),
-        (
-            ["a", "b", "c", "d", "e"],
-            "'a', 'b', 'c', 'd', and 'e'",
-        ),
-    ],
-)
-def test_format_component_list(components, expected):
-    assert dependency_mgmt._format_component_list(components) == expected
-
-
-@pytest.mark.parametrize(
-    "expected_aliases,expected_exception_msg",
-    [
-        (
-            {"other-component": "aliased", "third-component": "aliased"},
-            "Duplicate component alias 'aliased': components "
-            + "'other-component' and 'third-component' are aliased to 'aliased'",
-        ),
-        (
-            {"other-component": "third-component", "third-component": "aliased"},
-            "Component 'other-component' aliases existing component 'third-component'",
-        ),
-        (
-            {
-                "test-component": "third-component",
-                "other-component": "third-component",
-                "third-component": "aliased",
-            },
-            "Components 'other-component' and 'test-component' alias "
-            + "existing component 'third-component'",
-        ),
-    ],
-)
-@patch.object(dependency_mgmt, "kapitan_inventory")
-def test_discover_components_duplicate_aliases(
-    patch_inventory, data: Config, expected_aliases, expected_exception_msg
-):
-    _setup_mock_inventory(patch_inventory, expected_aliases)
-
-    with pytest.raises(KeyError) as e:
-        dependency_mgmt._discover_components(data)
-
-    assert e.value.args[0] == expected_exception_msg
 
 
 @patch("commodore.dependency_mgmt._read_components")
@@ -518,104 +200,6 @@ def test_fetch_components_is_minimal(
     for component in other_components:
         assert component not in data._components
         assert not (tmp_path / "dependencies" / component).exists()
-
-
-def test_write_jsonnetfile(data: Config, tmp_path: Path):
-    data.register_component(Component("test-component", work_dir=tmp_path))
-    data.register_component(Component("test-component-2", work_dir=tmp_path))
-    dirs = [
-        "dependencies/test-component",
-        "dependencies/test-component-2",
-        "dependencies/lib",
-    ]
-
-    file = tmp_path / "jsonnetfile.json"
-
-    dependency_mgmt.write_jsonnetfile(file, dependency_mgmt.jsonnet_dependencies(data))
-
-    with open(file) as jf:
-        jf_string = jf.read()
-        assert jf_string[-1] == "\n"
-        jf_contents = json.loads(jf_string)
-        assert jf_contents["version"] == 1
-        assert jf_contents["legacyImports"]
-        deps = jf_contents["dependencies"]
-        for dep in deps:
-            assert dep["source"]["local"]["directory"] in dirs
-
-
-def test_inject_essential_libraries(tmp_path: Path):
-    file = tmp_path / "jsonnetfile.json"
-    dependency_mgmt.write_jsonnetfile(file, [])
-
-    dependency_mgmt.inject_essential_libraries(file)
-
-    with open(file) as jf:
-        jf_string = jf.read()
-        assert jf_string[-1] == "\n"
-        jf_contents = json.loads(jf_string)
-        assert jf_contents["version"] == 1
-        assert jf_contents["legacyImports"]
-        deps = jf_contents["dependencies"]
-        assert len(deps) == 1
-        assert (
-            deps[0]["source"]["git"]["remote"]
-            == "https://github.com/bitnami-labs/kube-libsonnet"
-        )
-        assert deps[0]["version"] == "v1.19.0"
-
-
-def test_clear_jsonnet_lock_file(tmp_path: Path):
-    jsonnetfile = tmp_path / "jsonnetfile.json"
-    jsonnet_lock = tmp_path / "jsonnetfile.lock.json"
-    with open(jsonnetfile, "w") as jf:
-        json.dump(
-            {
-                "version": 1,
-                "dependencies": [
-                    {
-                        "source": {
-                            "git": {
-                                "remote": "https://github.com/brancz/kubernetes-grafana.git",
-                                "subdir": "grafana",
-                            }
-                        },
-                        "version": "master",
-                    }
-                ],
-                "legacyImports": True,
-            },
-            jf,
-        )
-    with open(jsonnet_lock, "w") as jl:
-        json.dump(
-            {
-                "version": 1,
-                "dependencies": [
-                    {
-                        "source": {
-                            "git": {
-                                "remote": "https://github.com/brancz/kubernetes-grafana.git",
-                                "subdir": "grafana",
-                            }
-                        },
-                        "version": "57b4365eacda291b82e0d55ba7eec573a8198dda",
-                        "sum": "92DWADwGjnCfpZaL7Q07C0GZayxBziGla/O03qWea34=",
-                    }
-                ],
-                "legacyImports": True,
-            },
-            jl,
-        )
-    dependency_mgmt.fetch_jsonnet_libraries(tmp_path)
-
-    assert jsonnet_lock.is_file()
-    with open(jsonnet_lock, "r") as file:
-        data = json.load(file)
-        assert (
-            data["dependencies"][0]["version"]
-            != "57b4365eacda291b82e0d55ba7eec573a8198dda"
-        )
 
 
 def _setup_register_components(tmp_path: Path):
